requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
demoji==1.1.0
filelock==3.3.1
huggingface-hub==0.0.19
idna==3.3
joblib==1.1.0
numpy==1.21.2
packaging==21.0
pandas==1.3.4
patricia-trie==10
PersianStemmer==1.0.1
pkg_resources==0.0.0
progressbar2==3.55.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
PyYAML==6.0
regex==2021.10.8
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tokenizers==0.10.3
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
